added support for get all and get by id

+++ src/main/java/com/pluresidea/api/emergencyincident/repository/IncidentRepository.java
@@ -3,7 +3,5 @@ package com.pluresidea.api.emergencyincident.repository;
 import com.pluresidea.api.emergencyincident.entity.Incident;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IncidentRepository extends JpaRepository<Incident, Long> {
-
-
+public interface IncidentRepository extends JpaRepository<Incident, Integer> {
 }
